third homework

+++ book-service/src/main/java/book/service/controller/BookController.java
@@ -1,11 +1,7 @@
 package book.service.controller;
 
 import book.service.controller.request.Request;
-import book.service.controller.response.ApiError;
 import book.service.controller.response.BookResponse;
-import book.service.entity.Author;
-import book.service.entity.Book;
-import book.service.service.AuthorService;
 import book.service.service.BookService;
 import jakarta.validation.Valid;
 import org.slf4j.Logger;
@@ -13,20 +9,26 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.UUID;
+
 
 @RestController
 @RequestMapping("/api/book")
 public class BookController {
   private final BookService bookService;
+  private final AuthorRegistryGateway authorRegistryGateway;
   private static final Logger logger = LoggerFactory.getLogger(BookController.class);
 
   @Autowired
-  public BookController(BookService bookService) {
+  public BookController(BookService bookService, AuthorRegistryGateway authorRegistryGateway) {
     this.bookService = bookService;
+    this.authorRegistryGateway = authorRegistryGateway;
   }
 
   @PostMapping("")
   public BookResponse createBook(@Valid @RequestBody Request.RequestToCreateBook request) {
+    String requestId = UUID.randomUUID().toString();
+    authorRegistryGateway.createBook(request.getFirstName(), request.getLastName(), request.getTitle(), requestId);
     return (bookService.createBook(request.getFirstName(), request.getLastName(), request.getTitle()));
   }
 
@@ -37,7 +39,10 @@ public class BookController {
 
   @DeleteMapping("/{id}")
   public void deleteBook(@PathVariable Long id) {
-    bookService.deleteBook(id);
+    String requestId = UUID.randomUUID().toString();
+    if (authorRegistryGateway.checkBook(bookService.getBookInfo(id), requestId)) {
+      bookService.deleteBook(id);
+    }
   }
 
 }
